@@ -1,4 +1,6 @@
 <script>
+  import { run } from 'svelte/legacy'
+
   import expand from '/static/expand.svg'
   import ason from '/static/css/ason.png'
   import { transformFilename } from '../../data.js'
@@ -6,13 +8,12 @@
   import { getSource } from '../../sources.js'
   import ModeToggle from './ModeToggle.svelte'
 
-  export let title, firstRegion, populationLabel, populationLink
+  let { title, firstRegion, populationLabel, populationLink, populationCount } =
+    $props()
 
   const source = getSource()
 
-  $: path = transformFilename(firstRegion)
-
-  let isControlsHidden = false
+  let isControlsHidden = $state(false)
 
   const triggerClose = () => {
     Dispatcher.setRegions([])
@@ -47,14 +48,13 @@
     const primary = Dispatcher.dataSegment.split('-')[1]
     setSegmentWithMode([newSegment, primary].join('-'), selection)
   }
-  let currentSegment = source.segments[0]
+  let currentSegment = $state(source.segments[0])
   const loadBlocks = () => {
     currentSegment = Dispatcher.dataSegment
   }
   Dispatcher.bind('load-blocks', loadBlocks)
 
-  let selection = []
-  $: setSegmentWithMode(Dispatcher.dataSegment, selection)
+  let selection = $state([])
   const setSegmentWithMode = (segment, selectedModes) => {
     let finalSegment = segment.split('-').slice(0, 2).join('-')
     if (selectedModes.length > 0) {
@@ -62,6 +62,10 @@
     }
     Dispatcher.setSegment(finalSegment)
   }
+  let path = $derived(transformFilename(firstRegion))
+  run(() => {
+    setSegmentWithMode(Dispatcher.dataSegment, selection)
+  })
 </script>
 
 <div class="nav-header" class:ason={source.brandingClass === 'ason'}>
@@ -76,12 +80,12 @@
       <button
         title="Learn More"
         class="btn-expand"
-        on:click={() =>
+        onclick={() =>
           document.getElementById('app').classList.toggle('map-view')}
       >
         <img alt="Toggle Details" src={expand} />
       </button>
-      <button title="Close Location" class="btn-close" on:click={triggerClose}>
+      <button title="Close Location" class="btn-close" onclick={triggerClose}>
         <svg
           xmlns="http://www.w3.org/2000/svg"
           height="24"
@@ -108,7 +112,7 @@
                 <a
                   href="#"
                   class="btn-segment"
-                  on:click={triggerSecondarySegment(control)}
+                  onclick={triggerSecondarySegment(control)}
                   class:selected={currentSegment.split('-')[0] ===
                     control.toLowerCase()}>{control}</a
                 >
@@ -121,16 +125,16 @@
         {#if populationLink}
           <a
             class="population-link"
-            href="https://www.stats.govt.nz/tools/2018-census-place-summaries/{path}#population-and-dwellings`"
+            href="https://www.stats.govt.nz/tools/2018-census-place-summaries/{path}#population-and-dwellings"
             target="_blank"
             rel="noopener noreferrer"
           >
             <strong class="population-label">{populationLabel}</strong>
-            <span class="population-count" />
+            <span class="population-count">{populationCount}</span>
           </a>
         {:else}
           <strong class="population-label">{populationLabel}</strong>
-          <span class="population-count" />
+          <span class="population-count">{populationCount}</span>
         {/if}
       </p>
     </div>
@@ -141,7 +145,7 @@
             <a
               href="#"
               class="btn-segment"
-              on:click={triggerSegment(control)}
+              onclick={triggerSegment(control)}
               class:selected={(source.detailsSecondaryControls != null
                 ? currentSegment.split('-')[1]
                 : currentSegment
