@@ -30,7 +30,7 @@
   </p>
 </div>
 <footer>
-  &copy; 2023 Turbomaps &middot; <a href="https://turbomaps.io">turbomaps.io</a>
+  &copy; 2025 Turbomaps &middot; <a href="https://turbomaps.io">turbomaps.io</a>
 </footer>
 
 <style>
