@@ -15,6 +15,7 @@ export const chooseBestName = (name, friendlyName) => {
 }
 
 export const humanRegionName = (nameArray, mode) => {
+  // if (nameArray === null) return ''
   const isZone = nameArray.length > 0 && !isNaN(parseInt(nameArray[0][0]))
   if (nameArray.length === 1) {
     return `${isZone ? 'Zone ' : ''}${nameArray[0]}`
@@ -29,8 +30,9 @@ export const humanRegionName = (nameArray, mode) => {
       .slice(0, -1)
       .join(', ')}, and ${nameArray.slice(-1)}`
   } else if (mode === 'title') {
-    return `${nameArray[0]} & ${nameArray.length - 1} other ${isZone ? 'zones' : 'areas'
-      }`
+    return `${nameArray[0]} & ${nameArray.length - 1} other ${
+      isZone ? 'zones' : 'areas'
+    }`
   }
   // should never get here
   return nameArray.join(', ')
