@@ -1,6 +1,8 @@
-import './components/map-tooltip.js'
+import { mount } from 'svelte'
 import App from './App.svelte'
 
-const app = new App({
+const app = mount(App, {
   target: document.getElementById('svelte-app'),
 })
+
+export default app
